Split search runtime modules
