+++ packages/worker/src/mcp/tools/search-response-size.ts
@@ -1,0 +1,43 @@
+import { charsPerToken, maxChars, maxTokens } from './search-constants.ts'
+
+export function truncateSearchText(text: string): string {
+	if (text.length <= maxChars) return text
+
+	return `${text.slice(0, maxChars)}\n\n--- TRUNCATED ---\nResponse was ~${Math.ceil(
+		text.length / charsPerToken,
+	).toLocaleString()} tokens (limit: ${maxTokens.toLocaleString()}). Lower the limit, maxResponseSize, or ask a shorter query.`
+}
+
+export function applyMaxResponseSize<TPayload>(
+	payload: TPayload,
+	maxResponseSize: number,
+	format: (payload: TPayload) => string,
+	trim: (payload: TPayload, count: number) => TPayload,
+	getCount: (payload: TPayload) => number,
+): { payload: TPayload; serialized: string } {
+	if (!Number.isFinite(maxResponseSize) || maxResponseSize <= 0) {
+		const serialized = format(payload)
+		return { payload, serialized }
+	}
+
+	const total = getCount(payload)
+	let low = 0
+	let high = total
+	let bestPayload = trim(payload, 0)
+	let bestSerialized = format(bestPayload)
+
+	while (low <= high) {
+		const mid = Math.floor((low + high) / 2)
+		const trimmedPayload = trim(payload, mid)
+		const serialized = format(trimmedPayload)
+		if (serialized.length <= maxResponseSize) {
+			bestPayload = trimmedPayload
+			bestSerialized = serialized
+			low = mid + 1
+		} else {
+			high = mid - 1
+		}
+	}
+
+	return { payload: bestPayload, serialized: bestSerialized }
+}
