+++ packages/worker/src/mcp/tools/search-tool-runner.ts
@@ -1,0 +1,597 @@
+import * as Sentry from '@sentry/cloudflare'
+import { resolvePublicUsername } from '#app/user-lookup.ts'
+import { type McpRegistrationAgent } from '#mcp/mcp-registration-agent.ts'
+import {
+	callerContextFields,
+	errorFields,
+	logMcpEvent,
+} from '#mcp/observability.ts'
+import { runPackageRetrievers } from '#worker/package-retrievers/service.ts'
+import { type RemoteConnectorStatus } from '#worker/remote-connector/status.ts'
+
+import { type MemoryToolSummary } from './memory-tool-context.ts'
+import {
+	escapeMarkdownText,
+	formatMarkdownInlineCode,
+} from './markdown-safety.ts'
+import { resolvePackageIdentitySearch } from './package-search-identity.ts'
+import {
+	defaultMaxResponseSize,
+	defaultSearchLimit,
+	maxChars,
+} from './search-constants.ts'
+import { buildExactPackageSearchResult, searchUnified } from './search-core.ts'
+import { resolveEntityDetail } from './search-detail.ts'
+import {
+	formatEntityDetailMarkdown,
+	formatSearchMarkdown,
+	type SearchEntityDetailStructured,
+	type SearchMatch,
+	type SearchResultStructuredContent,
+	toSlimStructuredMatches,
+} from './search-format.ts'
+import {
+	loadDownRemoteConnectorStatuses,
+	loadSearchRowsAndRegistry,
+	serializeRemoteConnectorStatus,
+} from './search-loaders.ts'
+import {
+	launchSearchMemoryEnrichment,
+	resolveSearchMemoryContext,
+	settleSearchMemoryEnrichment,
+} from './search-memory.ts'
+import {
+	applyMaxResponseSize,
+	truncateSearchText,
+} from './search-response-size.ts'
+import { elapsedMs } from './search-timing.ts'
+import {
+	type SearchPhaseTimings,
+	type SearchUnifiedResult,
+} from './search-types.ts'
+import { type SearchToolArgs } from './search-tool-definition.ts'
+import { resolveConversationId } from './tool-call-context.ts'
+import { prependToolMetadataContent } from './tool-response-content.ts'
+import { finishToolTiming, startToolTiming } from './tool-timing.ts'
+
+export async function runSearchTool(input: {
+	agent: McpRegistrationAgent
+	args: SearchToolArgs
+}) {
+	const { agent, args } = input
+	const timingStart = startToolTiming()
+	const conversationId = resolveConversationId(args.conversationId)
+	const callerContext = agent.getCallerContext()
+	const { baseUrl, hasUser } = callerContextFields(callerContext)
+	const userId = callerContext.user?.userId ?? null
+	const includeHiddenPackages = !!args.includeHiddenPackages
+	if (!args.query && !args.entity) {
+		const timing = finishToolTiming(timingStart)
+		logMcpEvent({
+			category: 'mcp',
+			tool: 'search',
+			toolName: 'search',
+			outcome: 'failure',
+			durationMs: timing.durationMs,
+			baseUrl,
+			hasUser,
+			sandboxError: false,
+			errorName: 'ValidationError',
+			errorMessage: 'Provide either "query" or "entity".',
+			message: 'Search request missing both query and entity.',
+			context: {
+				failurePhase: 'validation_error',
+			},
+		})
+		return {
+			content: prependToolMetadataContent(conversationId, [
+				{
+					type: 'text',
+					text: 'Error: Provide either "query" or "entity".',
+				},
+			]),
+			structuredContent: {
+				conversationId,
+				timing,
+				error: 'Provide either "query" or "entity".',
+			},
+			isError: true,
+		}
+	}
+	const limit = args.limit ?? defaultSearchLimit
+	const maxResponseSize = args.maxResponseSize ?? defaultMaxResponseSize
+	let warnings: Array<string> = []
+	let remoteConnectorDownStatuses: Array<RemoteConnectorStatus> = []
+	let username: string | null = null
+	const endToEndPhaseTimings: Partial<SearchPhaseTimings> = {}
+	let memoryEnrichmentPromise: Promise<MemoryToolSummary | null> =
+		Promise.resolve(null)
+	let memoryEnrichmentLaunchedAtMs: number | undefined
+
+	const searchSpan = async () => {
+		const query = args.query?.trim() ?? ''
+		username = await resolvePublicUsername({
+			db: agent.getEnv().APP_DB,
+			username: callerContext.user?.username ?? null,
+			email: callerContext.user?.email ?? null,
+		})
+		if (!args.entity) {
+			const launched = launchSearchMemoryEnrichment({
+				env: agent.getEnv(),
+				callerContext,
+				conversationId,
+				query: args.query,
+				memoryContext: args.memoryContext,
+			})
+			if (launched) {
+				memoryEnrichmentLaunchedAtMs = launched.launchedAtMs
+				memoryEnrichmentPromise = launched.promise
+			}
+		}
+		if (!args.entity && query) {
+			const identityResolution = await resolvePackageIdentitySearch({
+				db: agent.getEnv().APP_DB,
+				userId,
+				query,
+				baseUrl,
+				username,
+				includeHiddenPackages,
+			})
+			if (identityResolution.recognized) {
+				const remoteConnectorStatusStart = performance.now()
+				remoteConnectorDownStatuses = await loadDownRemoteConnectorStatuses({
+					env: agent.getEnv(),
+					callerContext,
+				})
+				endToEndPhaseTimings.remoteConnectorStatusMs = elapsedMs(
+					remoteConnectorStatusStart,
+				)
+				return {
+					mode: 'list' as const,
+					result: buildExactPackageSearchResult({
+						env: agent.getEnv(),
+						query,
+						match: identityResolution.match,
+					}),
+				}
+			}
+		}
+		const rowAndRegistryLoadStart = performance.now()
+		const rowsPromise = loadSearchRowsAndRegistry({
+			env: agent.getEnv(),
+			callerContext,
+			userId,
+			includeHiddenPackages,
+		}).then((rows) => {
+			endToEndPhaseTimings.rowAndRegistryLoadMs = elapsedMs(
+				rowAndRegistryLoadStart,
+			)
+			return rows
+		})
+		const retrieversStart = performance.now()
+		const retrieverRunPromise =
+			!args.entity && userId && query
+				? runPackageRetrievers({
+						env: agent.getEnv(),
+						baseUrl,
+						userId,
+						scope: 'search',
+						query,
+						includeHiddenPackages,
+						memoryContext: resolveSearchMemoryContext({
+							query,
+							memoryContext: args.memoryContext,
+						}),
+						conversationId,
+					}).then((retrieverRun) => {
+						endToEndPhaseTimings.retrieversMs = elapsedMs(retrieversStart)
+						return retrieverRun
+					})
+				: Promise.resolve({ results: [], warnings: [] }).then(
+						(retrieverRun) => {
+							endToEndPhaseTimings.retrieversMs = elapsedMs(retrieversStart)
+							return retrieverRun
+						},
+					)
+		const [searchRows] = await Promise.all([rowsPromise, retrieverRunPromise])
+		const remoteConnectorStatusStart = performance.now()
+		remoteConnectorDownStatuses = await loadDownRemoteConnectorStatuses({
+			env: agent.getEnv(),
+			callerContext,
+		})
+		endToEndPhaseTimings.remoteConnectorStatusMs = elapsedMs(
+			remoteConnectorStatusStart,
+		)
+		warnings = searchRows.warnings
+
+		if (args.entity) {
+			if (Array.isArray(args.entity)) {
+				const batchResults = await Promise.all(
+					args.entity.map(async (entityRef) => {
+						try {
+							const detail = await resolveEntityDetail({
+								agent,
+								callerContext,
+								userId,
+								username,
+								entity: entityRef,
+								searchRows,
+							})
+							return {
+								ok: true as const,
+								entityRef,
+								detail,
+							}
+						} catch (cause) {
+							const error =
+								cause instanceof Error ? cause : new Error(String(cause))
+							return {
+								ok: false as const,
+								entityRef,
+								error: error.message,
+							}
+						}
+					}),
+				)
+				return {
+					mode: 'entity-batch' as const,
+					results: batchResults,
+				}
+			}
+			return {
+				mode: 'entity' as const,
+				detail: await resolveEntityDetail({
+					agent,
+					callerContext,
+					userId,
+					username,
+					entity: args.entity,
+					searchRows,
+				}),
+			}
+		}
+
+		const retrieverRun = await retrieverRunPromise
+		warnings.push(...retrieverRun.warnings)
+		const result = await searchUnified({
+			env: agent.getEnv(),
+			query,
+			limit,
+			userId: userId ?? undefined,
+			registry: searchRows.registry,
+			optionalRows: searchRows,
+			retrieverResults: retrieverRun.results,
+		})
+
+		return {
+			mode: 'list' as const,
+			result,
+		}
+	}
+
+	try {
+		const outcome:
+			| {
+					mode: 'list'
+					result: SearchUnifiedResult
+			  }
+			| {
+					mode: 'entity'
+					detail: Awaited<ReturnType<typeof resolveEntityDetail>>
+			  }
+			| {
+					mode: 'entity-batch'
+					results: Array<
+						| {
+								ok: true
+								entityRef: string
+								detail: Awaited<ReturnType<typeof resolveEntityDetail>>
+						  }
+						| {
+								ok: false
+								entityRef: string
+								error: string
+						  }
+					>
+			  } = await Sentry.startSpan(
+			{
+				name: 'mcp.tool.search',
+				op: 'mcp.tool',
+				attributes: {
+					'mcp.tool': 'search',
+				},
+			},
+			searchSpan,
+		)
+
+		if (outcome.mode === 'entity') {
+			const entityResult = formatEntityDetailMarkdown(outcome.detail)
+			const timing = finishToolTiming(timingStart)
+			logMcpEvent({
+				category: 'mcp',
+				tool: 'search',
+				toolName: 'search',
+				outcome: 'success',
+				durationMs: timing.durationMs,
+				baseUrl,
+				hasUser,
+			})
+			return {
+				content: prependToolMetadataContent(conversationId, [
+					{
+						type: 'text',
+						text: truncateSearchText(entityResult.markdown),
+					},
+				]),
+				structuredContent: {
+					conversationId,
+					timing,
+					result: entityResult.structured,
+				},
+			}
+		}
+
+		if (outcome.mode === 'entity-batch') {
+			const timing = finishToolTiming(timingStart)
+			const structuredResults: Array<
+				SearchEntityDetailStructured | { entityRef: string; error: string }
+			> = []
+			const markdownParts: Array<string> = []
+			let successCount = 0
+			for (const entry of outcome.results) {
+				if (!entry.ok) {
+					structuredResults.push({
+						entityRef: entry.entityRef,
+						error: entry.error,
+					})
+					markdownParts.push(
+						`Error resolving ${formatMarkdownInlineCode(entry.entityRef)}: ${escapeMarkdownText(entry.error)}`,
+					)
+					continue
+				}
+				const entityResult = formatEntityDetailMarkdown(entry.detail)
+				const candidateStructured = [
+					...structuredResults,
+					entityResult.structured,
+				]
+				const hasFullDetail = structuredResults.some(
+					(result) => 'kind' in result && result.kind === 'entity',
+				)
+				if (
+					hasFullDetail &&
+					JSON.stringify(candidateStructured).length > maxChars
+				) {
+					const overflowError =
+						'Omitted from batch response (exceeds size budget). Look up individually with search({ entity }).'
+					structuredResults.push({
+						entityRef: entry.entityRef,
+						error: overflowError,
+					})
+					markdownParts.push(
+						`Error resolving ${formatMarkdownInlineCode(entry.entityRef)}: ${escapeMarkdownText(overflowError)}`,
+					)
+					continue
+				}
+				successCount += 1
+				structuredResults.push(entityResult.structured)
+				markdownParts.push(entityResult.markdown)
+			}
+			const allFailed = successCount === 0
+			logMcpEvent({
+				category: 'mcp',
+				tool: 'search',
+				toolName: 'search',
+				outcome: allFailed ? 'failure' : 'success',
+				durationMs: timing.durationMs,
+				baseUrl,
+				hasUser,
+				...(allFailed
+					? {
+							sandboxError: false,
+							errorName: 'EntityBatchError',
+							errorMessage: 'All entity lookups failed.',
+						}
+					: {}),
+			})
+			return {
+				content: prependToolMetadataContent(conversationId, [
+					{
+						type: 'text',
+						text: truncateSearchText(markdownParts.join('\n\n---\n\n')),
+					},
+				]),
+				structuredContent: {
+					conversationId,
+					timing,
+					result: structuredResults,
+					...(allFailed ? { error: 'All entity lookups failed.' } : {}),
+				},
+				...(allFailed ? { isError: true } : {}),
+			}
+		}
+
+		const normalizedRemoteConnectorStatuses =
+			remoteConnectorDownStatuses.length > 0
+				? remoteConnectorDownStatuses.map(serializeRemoteConnectorStatus)
+				: undefined
+		const memorySettlement = await settleSearchMemoryEnrichment({
+			env: agent.getEnv(),
+			callerContext,
+			conversationId,
+			promise: memoryEnrichmentPromise,
+			launchedAtMs: memoryEnrichmentLaunchedAtMs,
+		})
+		Object.assign(endToEndPhaseTimings, memorySettlement.phaseTimings)
+		warnings.push(...memorySettlement.warnings)
+		const searchMemories = memorySettlement.memories
+		const structuredWarnings = [...warnings]
+
+		const payload: {
+			matches: Array<SearchMatch>
+			offline: boolean
+			remoteConnectorStatuses?: Array<{
+				connectorId: string
+				state: string
+				connected: boolean
+				toolCount: number
+			}>
+		} = {
+			matches: outcome.result.matches,
+			offline: outcome.result.offline,
+			...(normalizedRemoteConnectorStatuses
+				? {
+						remoteConnectorStatuses: normalizedRemoteConnectorStatuses,
+					}
+				: {}),
+		}
+		const statefulAgent = agent as McpRegistrationAgent & {
+			state?: {
+				searchConversationIdsWithPreamble?: Array<string>
+			}
+			setState?: (state: {
+				searchConversationIdsWithPreamble?: Array<string>
+			}) => void
+		}
+		const searchConversationIdsWithPreamble = Array.isArray(
+			statefulAgent.state?.searchConversationIdsWithPreamble,
+		)
+			? (statefulAgent.state?.searchConversationIdsWithPreamble ?? [])
+			: []
+		const includePreamble =
+			!args.conversationId ||
+			!searchConversationIdsWithPreamble.includes(conversationId)
+		if (includePreamble && typeof statefulAgent.setState === 'function') {
+			statefulAgent.setState({
+				...(statefulAgent.state ?? {}),
+				searchConversationIdsWithPreamble: [
+					...searchConversationIdsWithPreamble,
+					conversationId,
+				],
+			})
+		}
+		const formattingStartMs = performance.now()
+		const { payload: trimmedPayload, serialized } = applyMaxResponseSize(
+			payload,
+			maxResponseSize,
+			(value) =>
+				formatSearchMarkdown({
+					matches: value.matches,
+					warningCount: structuredWarnings.length,
+					includePreamble,
+				}),
+			(value, count) => ({
+				...value,
+				matches: value.matches.slice(0, count),
+			}),
+			(value) => value.matches.length,
+		)
+		const trimmedMatchCount = Math.max(
+			0,
+			outcome.result.matches.length - trimmedPayload.matches.length,
+		)
+		const formattingMs = elapsedMs(formattingStartMs)
+		const result: SearchResultStructuredContent = {
+			offline: trimmedPayload.offline,
+			warnings: structuredWarnings,
+			...(outcome.result.guidance
+				? {
+						guidance: outcome.result.guidance,
+					}
+				: {}),
+			telemetry: {
+				...outcome.result.telemetry,
+				topResultTypes: trimmedPayload.matches
+					.slice(0, 5)
+					.map((match) => match.type),
+				trimmedMatchCount,
+				responseTrimmed: trimmedMatchCount > 0,
+			},
+			phaseTimings: {
+				...outcome.result.phaseTimings,
+				...endToEndPhaseTimings,
+				formattingMs,
+			},
+			...(searchMemories
+				? {
+						memories: searchMemories,
+					}
+				: {}),
+			...(trimmedPayload.remoteConnectorStatuses
+				? {
+						remoteConnectorStatuses: trimmedPayload.remoteConnectorStatuses,
+					}
+				: {}),
+			matches: toSlimStructuredMatches({
+				matches: trimmedPayload.matches,
+				baseUrl,
+				username,
+			}),
+		}
+		const timing = finishToolTiming(timingStart)
+		logMcpEvent({
+			category: 'mcp',
+			tool: 'search',
+			toolName: 'search',
+			outcome: 'success',
+			durationMs: timing.durationMs,
+			baseUrl,
+			hasUser,
+			message: 'Search completed successfully.',
+			context: {
+				task: outcome.result.intent.task.name,
+				intentConfidence: outcome.result.intent.confidence,
+				entityCount: outcome.result.intent.entities.length,
+				actionCount: outcome.result.intent.actions.length,
+				constraintCount: outcome.result.intent.constraints.length,
+				candidateCounts: outcome.result.telemetry.candidateCounts,
+				topResultTypes: result.telemetry?.topResultTypes ?? [],
+				responseTrimmed: result.telemetry?.responseTrimmed ?? false,
+				trimmedMatchCount,
+				offline: outcome.result.offline,
+				warningsCount: warnings.length,
+				phaseTimings: result.phaseTimings,
+			},
+		})
+		return {
+			content: prependToolMetadataContent(conversationId, [
+				{
+					type: 'text',
+					text: truncateSearchText(serialized),
+				},
+			]),
+			structuredContent: {
+				conversationId,
+				timing,
+				result,
+			},
+		}
+	} catch (cause) {
+		const timing = finishToolTiming(timingStart)
+		const error = cause instanceof Error ? cause : new Error(String(cause))
+		const { errorName, errorMessage } = errorFields(error)
+		logMcpEvent({
+			category: 'mcp',
+			tool: 'search',
+			toolName: 'search',
+			outcome: 'failure',
+			durationMs: timing.durationMs,
+			baseUrl,
+			hasUser,
+			sandboxError: false,
+			errorName,
+			errorMessage,
+			cause: error,
+		})
+		return {
+			content: prependToolMetadataContent(conversationId, [
+				{ type: 'text', text: `Error: ${error.message}` },
+			]),
+			structuredContent: {
+				conversationId,
+				timing,
+				error: error.message,
+			},
+			isError: true,
+		}
+	}
+}
